Rename random_circuit to random_graph
