+++ src/lib.rs
@@ -23,7 +23,7 @@ mod isomorphism;
 mod iterators;
 mod layout;
 mod matching;
-mod random_circuit;
+mod random_graph;
 mod shortest_path;
 mod transitivity;
 mod traversal;
@@ -36,7 +36,7 @@ use dag_algo::*;
 use isomorphism::*;
 use layout::*;
 use matching::*;
-use random_circuit::*;
+use random_graph::*;
 use shortest_path::*;
 use transitivity::*;
 use traversal::*;
